Make reality details render
