+++ ui/components/Realities/index.js
@@ -1,126 +1,51 @@
-import React, { useEffect } from "react";
-import { useQuery, useMutation, gql } from "@apollo/client";
+import React from "react";
+//import styled from "styled-components";
+import { Col, Container, Row } from "reactstrap";
 import HappySpinner from "components/HappySpinner";
-import createRealitiesApollo from "lib/realities/createRealitiesApollo";
-import { REALITIES_CREATE_SUBSCRIPTION } from "lib/realities/subscriptions";
+import SubMenu from "components/SubMenu";
+import NeedsContainer from "components/Realities/NeedsContainer";
+//import Search from 'components/Search';
+import DetailViewContainer from "components/Realities/Details/DetailViewContainer";
 
-const GET_NEEDS = gql`
-  query Needs {
-    needs {
-      nodeId
-      title
-      fulfilledBy {
-        nodeId
-        title
-        realizer {
-          nodeId
-          name
-        }
-      }
-    }
-  }
-`;
+//const SearchRow = styled(Row)`
+//  margin-bottom: 20px;
+//`;
 
-const CREATE_NEED = gql`
-  mutation CreateNeed($title: String!) {
-    createNeed(title: $title) {
-      nodeId
-      title
-      fulfilledBy {
-        nodeId
-        title
-        realizer {
-          nodeId
-          name
-        }
-      }
-    }
-  }
-`;
+const RealitiesHome = ({ currentOrgMember, currentUser }) => (
+  <>
+    <SubMenu currentOrgMember={currentOrgMember} />
+    <div className="page">
+      <Container fluid>
+        {/*<SearchRow className="d-md-none">
+      <Col>
+        <Search />
+      </Col>
+    </SearchRow>*/}
+        <Row>
+          <Col md="6" lg={{ size: 4, offset: 1 }}>
+            <NeedsContainer currentUser={currentUser} />
+          </Col>
+          <Col md="6">
+            <DetailViewContainer currentUser={currentUser} viewResp={false} />
+            <DetailViewContainer currentUser={currentUser} viewResp />
+          </Col>
+        </Row>
+      </Container>
+    </div>
+  </>
+);
 
-const realitiesApollo = createRealitiesApollo();
-
-const RealitiesPage = () => {
+const Realities = ({ currentOrgMember, currentUser }) => {
   const onServer = typeof window === "undefined";
-  const { data, error, loading, subscribeToMore } = useQuery(GET_NEEDS, {
-    skip: onServer,
-    client: realitiesApollo,
-  });
-  const [createNeed, { mutError }] = useMutation(CREATE_NEED, {
-    skip: onServer,
-    client: realitiesApollo,
-    update: (cache, { data: { createNeed: createNeedRes } }) => {
-      const { needs } = cache.readQuery({ query: GET_NEEDS });
-
-      const alreadyExists =
-        needs.filter((need) => need.nodeId === createNeedRes.nodeId).length > 0;
-      if (!alreadyExists) {
-        cache.writeQuery({
-          query: GET_NEEDS,
-          data: { needs: [createNeedRes, ...needs] },
-        });
-      }
-    },
-  });
-
-  useEffect(
-    () =>
-      subscribeToMore({
-        document: REALITIES_CREATE_SUBSCRIPTION,
-        updateQuery: (prev, { subscriptionData }) => {
-          if (!subscriptionData.data) return prev;
-          const { realityCreated } = subscriptionData.data;
-
-          if (realityCreated.__typename !== "Need") return prev;
-
-          const alreadyExists =
-            prev.needs.filter((need) => need.nodeId === realityCreated.nodeId)
-              .length > 0;
-
-          if (alreadyExists) return prev;
-          return { needs: [realityCreated, ...prev.needs] };
-        },
-      }),
-    [subscribeToMore]
-  );
-
-  if (error || mutError) {
-    console.error("realities error", error, mutError);
-    return "error";
-  }
-  if (loading) return <HappySpinner />;
+  // until we've fixed ssr for the realities apollo client we'll just skip ssr here
+  if (onServer) return <HappySpinner />;
 
   return (
-    <div>
-      <div>Needs & Resps:</div>
-      <ul className="list-disc">
-        {data?.needs.map((need) => {
-          return (
-            <React.Fragment key={need.nodeId}>
-              <li>{need.title}</li>
-              {need.fulfilledBy.map((resp) => {
-                return (
-                  <li key={resp.nodeId} className="ml-5">
-                    {resp.title}
-                  </li>
-                );
-              })}
-            </React.Fragment>
-          );
-        })}
-      </ul>
-      <button
-        className="mt-4 border-2 border-green-300"
-        onClick={() =>
-          createNeed({
-            variables: { title: `New need ${~~(Math.random() * 1000)}` },
-          })
-        }
-      >
-        Create new need
-      </button>
-    </div>
+    <RealitiesHome
+      currentOrgMember={currentOrgMember}
+      currentUser={currentUser}
+    />
   );
 };
 
-export default RealitiesPage;
+export default Realities;
