+++ ui/pages/realities/need/[needId].js
@@ -1,0 +1,3 @@
+import RealitiesPage from "..";
+
+export default RealitiesPage;
